moved all dto, response and request models, and entities to package "models"
