+++ backend/src/main/java/com/htecgroup/SkyNest/model/dto/UserDto.java
@@ -1,4 +1,4 @@
-package com.htecgroup.SkyNest.dto;
+package com.htecgroup.SkyNest.model.dto;
 
 import lombok.Data;
