+++ backend/src/main/java/com/htecgroup/SkyNest/model/request/UserLoginRequest.java
@@ -1,4 +1,4 @@
-package com.htecgroup.SkyNest.io.request;
+package com.htecgroup.SkyNest.model.request;
 
 import lombok.Data;
