+++ backend/src/main/java/com/htecgroup/SkyNest/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.htecgroup.SkyNest.repository;
 
-import com.htecgroup.SkyNest.enitity.UserEntity;
+import com.htecgroup.SkyNest.model.enitity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
