+++ backend/src/main/java/com/htecgroup/SkyNest/service/UserService.java
@@ -1,6 +1,6 @@
 package com.htecgroup.SkyNest.service;
 
-import com.htecgroup.SkyNest.dto.UserDto;
+import com.htecgroup.SkyNest.model.dto.UserDto;
 
 public interface UserService {
